# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Apex Arena - Salesforce coding practice API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ProblemCategory =
  (typeof ProblemCategory)[keyof typeof ProblemCategory];

export const ProblemCategory = {
  trigger: "trigger",
  async_apex: "async_apex",
  classes: "classes",
  soql: "soql",
} as const;

export type Difficulty = (typeof Difficulty)[keyof typeof Difficulty];

export const Difficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

/**
 * The shape of code expected from the user
 */
export type ProblemKind = (typeof ProblemKind)[keyof typeof ProblemKind];

export const ProblemKind = {
  trigger: "trigger",
  class: "class",
  batch: "batch",
  queueable: "queueable",
  schedulable: "schedulable",
  future: "future",
  soql: "soql",
} as const;

export type SubmissionStatus =
  (typeof SubmissionStatus)[keyof typeof SubmissionStatus];

export const SubmissionStatus = {
  accepted: "accepted",
  wrong_answer: "wrong_answer",
  runtime_error: "runtime_error",
  compile_error: "compile_error",
  time_limit_exceeded: "time_limit_exceeded",
  governor_limit_exceeded: "governor_limit_exceeded",
} as const;

export interface TagCount {
  tag: string;
  count: number;
}

export interface ProblemSummary {
  id: number;
  slug: string;
  title: string;
  category: ProblemCategory;
  difficulty: Difficulty;
  kind: ProblemKind;
  tags: string[];
  /** Percentage 0-100 */
  acceptanceRate: number;
  totalSubmissions: number;
  /** Whether the requesting user has solved it (only present when userId is supplied) */
  solved?: boolean;
  /** Whether this problem requires a Pro subscription */
  isPremium?: boolean;
}

export interface TestCase {
  name: string;
  description: string;
  hidden: boolean;
}

export interface ProblemDetail {
  id: number;
  slug: string;
  title: string;
  category: ProblemCategory;
  difficulty: Difficulty;
  kind: ProblemKind;
  tags: string[];
  /** Markdown-style problem description */
  statement: string;
  hints: string[];
  starterCode: string;
  sampleTests: TestCase[];
  hiddenTestCount: number;
  acceptanceRate: number;
  totalSubmissions: number;
  solved?: boolean;
  /** Whether this problem requires a Pro subscription */
  isPremium?: boolean;
}

export interface TestResult {
  name: string;
  passed: boolean;
  message: string;
  hidden: boolean;
  executionTimeMs: number;
}

export interface GovernorLimits {
  soqlQueries: number;
  soqlQueryRows: number;
  dmlStatements: number;
  dmlRows: number;
  cpuTimeMs: number;
  heapSizeBytes: number;
}

export interface SubmissionResult {
  id: number;
  status: SubmissionStatus;
  passedCount: number;
  totalCount: number;
  results: TestResult[];
  debugLog: string[];
  compileError?: string | null;
  runtimeError?: string | null;
  governorLimits: GovernorLimits;
  createdAt: string;
}

export interface RunResult {
  results: TestResult[];
  debugLog: string[];
  compileError?: string | null;
  runtimeError?: string | null;
  governorLimits: GovernorLimits;
}

export interface SubmissionSummary {
  id: number;
  problemSlug: string;
  problemTitle: string;
  category: ProblemCategory;
  difficulty: Difficulty;
  status: SubmissionStatus;
  passedCount: number;
  totalCount: number;
  createdAt: string;
}

export interface SubmissionInput {
  userId: string;
  problemSlug: string;
  code: string;
}

export interface RunCodeInput {
  problemSlug: string;
  code: string;
}

export interface UserProfile {
  id: string;
  username: string;
  joinedAt: string;
  solvedCount: number;
  submissionCount: number;
  isSubscribed: boolean;
}

export interface UpsertUserInput {
  id: string;
  username: string;
}

export interface CategoryProgress {
  category: ProblemCategory;
  solved: number;
  total: number;
}

export interface DifficultyProgress {
  difficulty: Difficulty;
  solved: number;
  total: number;
}

export interface RecentActivity {
  problemSlug: string;
  problemTitle: string;
  status: SubmissionStatus;
  createdAt: string;
}

export interface UserStats {
  userId: string;
  username: string;
  solvedCount: number;
  submissionCount: number;
  acceptanceRate: number;
  currentStreakDays: number;
  byCategory: CategoryProgress[];
  byDifficulty: DifficultyProgress[];
  recentActivity: RecentActivity[];
}

export type PlatformStatsByCategoryItem = {
  category: ProblemCategory;
  count: number;
};

export type PlatformStatsByDifficultyItem = {
  difficulty: Difficulty;
  count: number;
};

export interface PlatformStats {
  totalProblems: number;
  totalSubmissions: number;
  totalUsers: number;
  byCategory: PlatformStatsByCategoryItem[];
  byDifficulty: PlatformStatsByDifficultyItem[];
}

export interface LeaderboardEntry {
  userId: string;
  username: string;
  solvedCount: number;
  submissionCount: number;
}

export type ListProblemsParams = {
  category?: ProblemCategory;
  difficulty?: Difficulty;
  tag?: string;
  search?: string;
  /**
   * If supplied, includes the user's solved status per problem
   */
  userId?: string;
};